token read done
